fix: Estate Planning Gifting Strategy tab - Add cash asset type support and fix type casting
## Gifting Strategy Bug Fixes (October 27, 2025)

### Fix 1: Unhandled 'cash' Asset Type in Liquidity Analyzer
- **Issue**: Personalized gifting strategy crashed with "Unhandled match case 'cash'"
- **Root Cause**: AssetLiquidityAnalyzer match statement didn't handle 'cash' asset type
- **Solution**: Added 'cash' case to match statement, classified as liquid asset
- **File**: app/Services/Estate/AssetLiquidityAnalyzer.php (lines 113-126)

### Fix 2: Type Error in Income-Based Gifting Strategy
- **Issue**: round() function error - string passed instead of float
- **Root Cause**: Database DECIMAL fields return strings in MySQL/Laravel
- **Solution**: Added explicit (float) casting for all income/expenditure values
- **File**: app/Services/Estate/PersonalizedGiftingStrategyService.php (lines 344-351)

## Other Changes Committed (October 27, 2025)

### Country Selector Implementation
- Added country tracking to all asset and liability forms (onboarding & Net Worth)
- CountrySelector component added to properties, investments, cash, liabilities
- Backend saves country field for all asset/liability types
- Moved PropertyForm country selector to Step 1 for better UX

### Emergency Fund & Expenditure Integration
- Fixed expenditure data flow from User Profile to Emergency Fund tab
- Auto-create expenditure_profile when user updates expenditure
- Changed ExpenditureProfile casts from decimal:2 to float (prevent NaN)
- Emergency Fund now displays total monthly expenditure correctly
- Fixed router link from /user-profile to /profile

### Database Migrations
- 2025_10_27_090614_add_country_to_properties_table.php
- 2025_10_27_090642_add_country_to_investment_accounts_table.php
- 2025_10_27_090643_add_country_to_savings_accounts_table.php
- 2025_10_27_090644_add_country_to_business_interests_table.php
- 2025_10_27_090645_add_country_to_chattels_table.php
- 2025_10_27_090647_add_country_to_cash_accounts_table.php
- 2025_10_27_090647_add_country_to_mortgages_table.php
- 2025_10_27_090648_add_country_to_liabilities_table.php

### Documentation Updates
- Updated bugsFixes2710.md with all fixes from October 27 session
- Updated OCTOBER_2025_FEATURES_UPDATE.md with bug fixes section
- Version updated to v0.1.2.4

## Impact
- ✅ Gifting Strategy tab now loads successfully with all asset types
- ✅ Cash accounts properly classified as immediately giftable liquid assets
- ✅ Income-based gifting strategies calculate correctly
- ✅ Type-safe calculations prevent PHP 8+ strict type errors
- ✅ Country tracking implemented across all asset and liability forms
- ✅ Emergency Fund displays expenditure data correctly

## Testing
- Manual testing completed - all features working as expected
- Estate Planning > Gifting Strategy tab loads without errors
- Personalized gifting strategy shows liquidity analysis
- Country selectors appear in all relevant forms
- Emergency Fund tab displays expenditure correctly

+++ app/Services/Estate/FutureValueCalculator.php
@@ -243,7 +243,7 @@ class FutureValueCalculator
         $totalFutureValue = 0;
 
         foreach ($assets as $asset) {
-            $currentValue = $asset->current_value ?? 0;
+            $currentValue = (float) ($asset->current_value ?? 0);
             $assetType = $asset->asset_type ?? 'other';
 
             // Get growth rate for this asset type, or use default
